+++ tools/shader_compiler/LLVM/test_shaders/std2/sampler.hpp
@@ -6,7 +6,7 @@ namespace skr::shader {
 struct [[builtin("sampler")]] Sampler 
 {
     template <typename T, uint32 Flags>
-    [[callop("SAMPLE2D")]] vec<T, 4> Sample(Image<T, Flags>& image, uint2 uv);
+    [[callop("SAMPLE2D")]] vec<T, 4> Sample(Image<T, Flags>& image, float2 uv);
 };
 
 } // namespace skr::shader
